[FEAT] - 테스트 코드 환경 구축
