--- src/test/java/com/bitbox/chatting/ChattingApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bitbox.chatting;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChattingApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
