+++ src/test/java/com/bitbox/chatting/service/ChattingServiceTest.java
@@ -1,0 +1,7 @@
+package com.bitbox.chatting.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ChattingServiceTest {
+
+}
